Removing Template folder, adding PDF Proof
When initially developing this book layout, it was thought that a
template might allow the creation of pages with specific layout -
however this is not the purpose of the template - which exists as a
template for a document as a whole.  Master pages were also experimented
with, and they proved equally unusable.